Move terminal output generator to own module.
output_generators/
	-Create a directory to contain all output generator modules for
	test-suite results.

test.js, output_generators/terminal.js
	-Move all terminal output generation code to own module,
	to reinforce the separation of test-running/results collection
	and output generation.
